Irix 6 hack.
svn:revision:9891
cvs:account:mike
cvs:branch:trunk

+++ rt/view.c
@@ -112,9 +112,9 @@ static struct scanline {
 
 /* Viewing module specific "set" variables */
 struct structparse view_parse[] = {
-	{"%d",	1, "bounces",	(int)&max_bounces,		FUNC_NULL },
-	{"%d",	1, "ireflect",	(int)&max_ireflect,		FUNC_NULL },
-	{"%f", ELEMENTS_PER_VECT, "background",(int)background,	FUNC_NULL },
+	{"%d",	1, "bounces",	(long)&max_bounces,		FUNC_NULL },
+	{"%d",	1, "ireflect",	(long)&max_ireflect,		FUNC_NULL },
+	{"%f", ELEMENTS_PER_VECT, "background",(long)background,	FUNC_NULL },
 	{"",	0, (char *)0,	0,				FUNC_NULL }
 };
